Makes IgnoreNodeFilter a separate class.
