+++ core/src/main/java/com/graphhopper/routing/ch/NodeContractor.java
@@ -349,29 +349,6 @@ class NodeContractor {
         return (int) meanDegree * 100;
     }
 
-    static class IgnoreNodeFilter implements EdgeFilter {
-        int avoidNode;
-        int maxLevel;
-        CHGraph graph;
-
-        IgnoreNodeFilter(CHGraph chGraph, int maxLevel) {
-            this.graph = chGraph;
-            this.maxLevel = maxLevel;
-        }
-
-        IgnoreNodeFilter setAvoidNode(int node) {
-            this.avoidNode = node;
-            return this;
-        }
-
-        @Override
-        public final boolean accept(EdgeIteratorState iter) {
-            // ignore if it is skipNode or adjNode is already contracted
-            int node = iter.getAdjNode();
-            return avoidNode != node && graph.getLevel(node) == maxLevel;
-        }
-    }
-
     static class Shortcut {
         int from;
         int to;
